+++ example/assets/js/main.js
@@ -1,0 +1,4 @@
+function add(a, b) {
+  return a + b;
+}
+console.log(add(10, 20));