+++ nodejs_server_test_auth_basic_and_ssl/public/js/main.js
@@ -1,0 +1,1 @@
+document.querySelector("#test").innerHTML = "this is a test";