+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.1.0
+
+- Added `pause` and `resume` methods for Android.
+- Added `pauseTimers` and `resumeTimers` methods.
+- Added new `historyUrl` optional parameter for `loadData` and `openData` methods and `InAppWebViewInitialData` class. It is used only on Android.
+- Fix "problems with onReceivedHttpAuthRequest when initialData is used" [#201](https://github.com/pichillilorenzo/flutter_inappwebview/issues/201)
+- Fix "System ui (status bar and navigation bar) doesn't hide automatically" [#202](https://github.com/pichillilorenzo/flutter_inappwebview/issues/202)
+
 ## 2.0.1+1
 
 - Fixed error "java.lang.ClassCastException: $Proxy1 cannot be cast to android.view.WindowManagerImpl" on Android when using native alert dialogs